+++ game/motd.go
@@ -3,36 +3,8 @@ package game
 import (
 	"fmt"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/dynamodb"
-	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
 	"github.com/google/uuid"
 )
-
-func (k *KeyPair) GetAllMOTDs() ([]*MOTD, error) {
-	input := &dynamodb.ScanInput{
-		TableName:        aws.String("motd"),
-		FilterExpression: aws.String("active = :active"),
-		ExpressionAttributeValues: map[string]*dynamodb.AttributeValue{
-			":active": {
-				BOOL: aws.Bool(true),
-			},
-		},
-	}
-
-	result, err := k.db.Scan(input)
-	if err != nil {
-		return nil, fmt.Errorf("error scanning MOTDs: %w", err)
-	}
-
-	var motds []*MOTD
-	err = dynamodbattribute.UnmarshalListOfMaps(result.Items, &motds)
-	if err != nil {
-		return nil, fmt.Errorf("error unmarshalling MOTDs: %w", err)
-	}
-
-	return motds, nil
-}
 
 // DisplayUnseenMOTDs shows unseen Messages of the Day to the player
 func DisplayUnseenMOTDs(server *Server, player *Player) error {
